Add Json Mappers from Podman Inspect to Docker Inspect
Added two mappers to fix a parsing issue where podman reports both the container state and the container entrypoint differently than docker.

The entrypoint issue is patched by converting the string into an array, the healthcheck has to be "faked" as podman doesn't require a healthcheck, meaning it can be blank. Need to check if podman statuses are lowercase...